refactor: removing redundant code
