+++ app/(main)/documents/[id]/page.tsx
@@ -21,7 +21,7 @@ import { Card, CardContent } from "@/components/ui/card";
 import { Separator } from "@/components/ui/separator";
 import { Skeleton } from "@/components/ui/skeleton";
 
-interface Document {
+export interface Document {
   id: string;
   title: string;
   description: string | null;
